Rearrange modules to fit 3 child system
