+++ src/scripts/app/cms/activities.cms.base.html
