--- src/scripts/app/cms/activities/config.js
@@ -1,19 +1,0 @@
-'use strict';
-module.exports =
-/*@ngInject*/
-function configure ($stateProvider) {
-  $stateProvider
-  .state('cms-activities', {
-    parent: 'cms-activities-base',
-    templateUrl: 'activities.cms.html',
-    contoller: 'ActivitiesCmsCtrl',
-    url: ''
-  })
-  .state('cms-activities-base', {
-    parent: 'cms',
-    templateUrl: 'activities.cms.base.html',
-    controller: 'ActivitiesCmsCtrl',
-    abstract: true,
-    url: '/activities'
-  });
-};
