--- src/scripts/app/cms/activities/practiceQuestions/config.js
@@ -1,12 +1,0 @@
-'use strict';
-module.exports =
-/*@ngInject*/
-function configure ($stateProvider) {
-  $stateProvider
-  .state('cms-activities-practice-questions', {
-    parent: 'cms-activities-base',
-    templateUrl: 'practiceQuestions.cms.html',
-    contoller: 'PracticeQuestionsCmsCtrl',
-    url: '/practiceQuestions'
-  });
-};
