--- src/scripts/app/cms/activities/practiceQuestions/controller.js
@@ -1,9 +1,0 @@
-'use strict';
-
-module.exports =
-
-/*@ngInject*/
-function PracticeQuestionsCmsCtrl (
-
-) {
-};
