--- src/scripts/app/cms/activities/practiceQuestions/module.js
@@ -1,10 +1,0 @@
-'use strict';
-
-module.exports =
-
-angular.module('quill-grammar.cms.activities.practiceQuestions', [
-
-])
-.config(require('./config.js'))
-.controller('PracticeQuestionsCmsCtrl', require('./controller.js'))
-;
