--- src/scripts/app/cms/activities/practiceQuestions/practiceQuestions.cms.html
@@ -1,1 +1,0 @@
-<h1>Practice Questions</h1>
