--- src/scripts/app/cms/activities/stories/module.js
@@ -1,10 +1,0 @@
-'use strict';
-
-module.exports =
-
-angular.module('quill-grammar.cms.activities.stories', [
-
-])
-.config(require('./config.js'))
-.controller('StoriesCmsCtrl', require('./controller.js'))
-;
