--- src/scripts/app/cms/activities/stories/stories.cms.html
@@ -1,1 +1,0 @@
-<h1>Stories</h1>
