+++ src/scripts/app/cms/cms.html
@@ -1,6 +1,7 @@
 <div class="cms">
-  <h1>Admin CMS</h1>
-  <a ui-sref="cms-activities()">Activities</a>
+  <h1>Create Grammar Materials</h1>
+  <a ui-sref="cms-grammar-activities()">Create Sentence Writing Activities</a>
+  <a ui-sref="cms-proofreading-activities()">Create Proofreading Activities</a>
   <a ui-sref="cms-concepts()">Concepts</a>
 </div>
