+++ src/scripts/app/cms/cms.module.js
@@ -4,7 +4,8 @@ module.exports =
 angular
   .module('quill-grammar.cms', [
     require('./concepts/module.js').name,
-    require('./activities/module.js').name,
+    require('./grammarActivities/module.js').name,
+    require('./proofreadingActivities/module.js').name,
   ])
   .config(require('./cms.config.js'))
   .controller('cms', require('./cms.controller.js'));
