+++ src/scripts/app/cms/grammarActivities/controller.js
@@ -3,7 +3,7 @@
 module.exports =
 
 /*@ngInject*/
-function ActivitiesCmsCtrl (
+function GrammarActivitiesCmsCtrl (
 
 ) {
 };
