+++ src/scripts/app/cms/grammarActivities/module.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports =
+
+angular.module('quill-grammar.cms.grammarActivities', [
+
+])
+.config(require('./config.js'))
+.controller('GrammarActivitiesCmsCtrl', require('./controller.js'))
+;
