+++ src/scripts/app/cms/proofreadingActivities/config.js
@@ -1,0 +1,12 @@
+'use strict';
+module.exports =
+/*@ngInject*/
+function configure ($stateProvider) {
+  $stateProvider
+  .state('cms-proofreading-activities', {
+    parent: 'cms-activities-base',
+    templateUrl: 'proofreading.activities.cms.html',
+    contoller: 'ProofreadingActivitiesCmsCtrl',
+    url: '/proofreadingActivities'
+  });
+};
