+++ src/scripts/app/cms/proofreadingActivities/controller.js
@@ -3,7 +3,7 @@
 module.exports =
 
 /*@ngInject*/
-function StoriesCmsCtrl (
+function ProofreadingActivitiesCmsCtrl (
 
 ) {
 };
