+++ src/scripts/app/cms/proofreadingActivities/proofreading.activities.cms.html
@@ -1,0 +1,1 @@
+<h3>Proofreading Activities</h3>
